refactor: move MixinSrcUrl to mixin_src_url.rs
