+++ components/restreamer/src/state/mixin.rs
@@ -1,4 +1,10 @@
+#![allow(clippy::module_name_repetitions)]
+
 mod delay;
+mod mixin_src_url;
 mod volume;
 
-pub use self::{delay::Delay, volume::Volume, volume::VolumeLevel};
+pub use self::{
+    delay::Delay, mixin_src_url::MixinSrcUrl, volume::Volume,
+    volume::VolumeLevel,
+};
